changed the data type to record temperature and humidity values as float instead of int (To support with DHT22 modules)

+++ Car_in_garage_detector1/Car_in_garage_detector1.ino
@@ -42,8 +42,8 @@ const int msgBodyBytes = nrfPayloadBytesLimit - reservedBytes;
 struct GarageData {
   float carDistance;
   GarageDoor door;
-  int temperature;
-  int humidity;
+  float temperature;
+  float humidity;
   GarageLights lights;
   int airQuality;
   float batteryLevel;
